+++ tests/test_utils.py
@@ -1,4 +1,4 @@
-from scripts.optim.utils import (
+from scripts.optim_routes.utils import (
     Point,
     calculate_distances_dict,
     calculate_path_len,
